Implementation of new payment method for EPiServer #1 Instantiating of the checkout process, calculating discounts, etc. #2 Accepting and Error handlers #3 Capturing/Voiding the transaction #4

+++ Dintero.Checkout.Episerver/Models/DinteroOrderLine.cs
@@ -26,11 +26,11 @@ namespace Dintero.Checkout.Episerver.Models
 
 
         [JsonProperty(PropertyName = "amount")]
-        public decimal Amount { get; set; }
+        public int Amount { get; set; }
 
 
         [JsonProperty(PropertyName = "vat_amount")]
-        public decimal VatAmount { get; set; }
+        public int VatAmount { get; set; }
 
 
         [JsonProperty(PropertyName = "vat")]
